Add files via upload
The source code for the Good Vibration Multi Band. 
This code is programmed for device 1. Remember to change the variables for 'id' and 'idOther' for the 2nd device.
Remember also to type in your own Wifi name and password.
The code is made for a device which uses motor B. If you use motor A or both, then remember to uncomment them in the code.